# lxdapi/docs/swagger.yaml
definitions:
...
  lxdapi_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  models.CreateContainerRequest:
    properties:
      allow_nesting:
        type: boolean
      cpu:
        type: integer
      cpu_allowance:
        type: integer
      disk:
        type: integer
      egress:
        type: integer
      image:
        type: string
      ingress:
        type: integer
      io_read:
        type: integer
      io_write:
        type: integer
      ipv4_mapping_limit:
        type: integer
      ipv4_pool_limit:
        type: integer
      ipv6_mapping_limit:
        type: integer
      ipv6_pool_limit:
        type: integer
      memory:
        type: integer
      memory_swap:
        type: boolean
      name:
        type: string
      password:
        type: string
      privileged:
        type: boolean
      processes_limit:
        type: integer
      remark:
        type: string
      reverse_proxy_limit:
        type: integer
      traffic_limit:
        type: integer
      username:
        type: string
    required:
    - image
    - name
    type: object
  models.UpdateContainerConfigRequest:
    properties:
      allow_nesting:
        type: boolean
      cpu:
        type: integer
      cpu_allowance:
        type: integer
      disk:
        type: integer
      egress:
        type: integer
      ingress:
        type: integer
      io_read:
        type: integer
      io_write:
        type: integer
      ipv4_mapping_limit:
        type: integer
      ipv4_pool_limit:
        type: integer
      ipv6_mapping_limit:
        type: integer
      ipv6_pool_limit:
        type: integer
      memory:
        type: integer
      memory_swap:
        type: boolean
      privileged:
        type: boolean
      processes_limit:
        type: integer
      remark:
        type: string
      reverse_proxy_limit:
        type: integer
      traffic_limit:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/system/containers:
    get:
      consumes:
      - application/json
      description: 获取所有容器的列表信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取容器列表
      tags:
      - System API - 容器管理
    post:
      consumes:
      - application/json
      description: 创建一个新的LXD容器，支持配置CPU、内存、磁盘、带宽等资源
      parameters:
      - description: 容器创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateContainerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回任务ID
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建容器
      tags:
      - System API - 容器管理
  /api/system/containers/{name}:
    delete:
      consumes:
      - application/json
      description: 删除指定名称的容器及相关数据（危险操作）
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除任务已创建
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少容器名称
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 创建任务失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除容器
      tags:
      - System API - 容器管理
    get:
      consumes:
      - application/json
      description: 获取指定容器的详细信息和运行状态
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少容器名称
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取容器详情
      tags:
      - System API - 容器管理
  /api/system/containers/{name}/action:
    post:
      consumes:
      - application/json
      description: '对容器执行操作: start/stop/restart/pause/resume/reinstall/reset-password'
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: '操作类型: start/stop/restart/pause/resume/reinstall/reset-password'
        in: query
        name: action
        required: true
        type: string
      - description: 操作参数（reinstall需要image/password，reset-password需要password）
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 操作失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 容器操作
      tags:
      - System API - 容器管理
  /api/system/containers/{name}/config:
    put:
      consumes:
      - application/json
      description: 热更新容器配置，支持CPU、内存、磁盘、带宽等升降级，磁盘只能扩容
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 配置更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateContainerConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新容器配置
      tags:
      - System API - 容器管理
...
swagger: "2.0"